+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## dev
 
+* added User, UserAuthentication and Users to domain layer
 * added link for HAL and HTML
 * initial version